requirements.txt
Flask==2.3.3
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.2.0
pyheif==0.7.1
pillow-heif==0.12.0
